Create karli.md
Hi everyone!

My name is Karli Snyder and I am a history major, minoring in Medieval Studies and Art.  This is my last year as an undergrad and I hope to begin an MA in Museum Studies next Fall.  I am particularly interested in Medieval and Early Modern history, and my MA studies will focus on the use of technology in the modern museum setting (digital archives and online collections).

When I am not in class, I enjoy being lazy at home with my husband and our two dogs.  We are runners and so are our pups.  I also enjoy yarn arts (i.e., knitting, crochet, cross stitch, embroidery) and often sell one-of-a-kind pieces of clothing that I hand-make.