I think the thing that intrigued me most about History is the idea of being able to walk in someone else’s shoes who has already experienced life or had lived through a particularly interesting part of time be it considered good or difficult. It’s interesting to see how the person made choices and how much their choices were influenced by circumstances that were not always in their control. I love to see how certain things played out with regards to History and I like to play devil’s advocate with History sometimes like what if so and so made this decision or if they had access to this resource how would things have ended up differently or would they have played more or less the same way. History feels alive when you look at it this way it feels intuitive and hands on when you can look at historical documents or other types of primary sources that you are researching and figure out what the person you are studying is thinking is the person who’s writing the primary source was affected by events that were unfolding around them in real time. It’s like you are a detective trying to gather facts and evidence in order to get be able to create a clearer picture of an event or person from the past.  The professed purpose that resonates with me the most were theological confirmations and questionings and myth breaking. The first professed purpose  made me think of when I growing up in a very religious household when I was younger and most of the time religion leaves very little room for historical context or the ability to ask questions about certain aspects of religion such as the story of creation itself. When dealing with religion it seems that its more concerned with keeping the original belief and constructing and searching for any evidence to support the belief instead of letting the evidence guide you to where you should go and make corrections and adjustments to said belief. I liked the idea that Beverly puts here when he said, “That is what makes challenging such foundation myths so important, and that’s what reveals another vital role for historians as non-conformist irritants.” (Southgate,42) I also liked when he was discussing myth breaking when it came to important historical events like the holocaust or 9/11 and how certain people who traffic in conspiracy and denialism or these events need to be put in check and for the victims of these atrocities be remembered and given a voice.  The professed purpose that resonated with me the least was probably not that I hated it or disliked it but wished there was more on the subject was cultivation. From what I’m assessing is that cultivation was something meant for wealthy or upper-class people to have to have the appreciation of History and knowledge of historical facts as a way to make you a better or more learned person. It just seemed kind of classist to me or outdated, but where I liked the notion was where Southgate was talking about Lord Bolingbrook and how he was making the point about being less ignorant to other people’s history’s and experiences to not be tied to a nationalist ideal of who you are. I found this section very interesting and wished Southgate could have added more to the professed purpose of cultivation.
I think the thing that intrigued me most about History is the idea of being able to walk in someone else’s shoes who has already experienced life or had lived through a particularly interesting part of time be it considered good or difficult. It’s interesting to see how the person made choices and how much their choices were influenced by circumstances that were not always in their control. I love to see how certain things played out with regards to History and I like to play devil’s advocate with History sometimes like what if so and so made this decision or if they had access to this resource how would things have ended up differently or would they have played more or less the same way. History feels alive when you look at it this way it feels intuitive and hands on when you can look at historical documents or other types of primary sources that you are researching and figure out what the person you are studying is thinking is the person who’s writing the primary source was affected by events that were unfolding around them in real time. It’s like you are a detective trying to gather facts and evidence in order to get be able to create a clearer picture of an event or person from the past. 
The professed purpose that resonates with me the most were theological confirmations and questionings and myth breaking. The first professed purpose  made me think of when I growing up in a very religious household when I was younger and most of the time religion leaves very little room for historical context or the ability to ask questions about certain aspects of religion such as the story of creation itself. When dealing with religion it seems that its more concerned with keeping the original belief and constructing and searching for any evidence to support the belief instead of letting the evidence guide you to where you should go and make corrections and adjustments to said belief. I liked the idea that Beverly puts here when he said, “That is what makes challenging such foundation myths so important, and that’s what reveals another vital role for historians as non-conformist irritants.” (Southgate,42) I also liked when he was discussing myth breaking when it came to important historical events like the holocaust or 9/11 and how certain people who traffic in conspiracy and denialism or these events need to be put in check and for the victims of these atrocities be remembered and given a voice. 
The professed purpose that resonated with me the least was probably not that I hated it or disliked it but wished there was more on the subject was cultivation. From what I’m assessing is that cultivation was something meant for wealthy or upper-class people to have to have the appreciation of History and knowledge of historical facts as a way to make you a better or more learned person. It just seemed kind of classist to me or outdated, but where I liked the notion was where Southgate was talking about Lord Bolingbrook and how he was making the point about being less ignorant to other people’s history’s and experiences to not be tied to a nationalist ideal of who you are. I found this section very interesting and wished Southgate could have added more to the professed purpose of cultivation.