michael m - enlightenment history
The characterization of the Enlightenment as “man’s release from his self-imposed tutelage” by Kant hints at the heart of this social movement as a turning away from blind submission to authority and faith in favor of individual human reason and liberty. This entailed a sharp criticism of traditional religious interpretations of the world, as exemplified particularly by Voltaire’s famous attacks on religion and Gibbon’s criticism of Christianity as contributing to the social decay and ultimate downfall of the Roman Empire. It was also a movement concerned with nation-building and national consciousness, which interacted varyingly with post-Reformation religious thought, both syncretically and critically.

History was appropriated by some Enlightenment writers as a means by which to derive a scientific understanding of human social interaction and provide a template from which to improve human society. It was also used to justify rising nationalist sentiment and revolutionary Republicanism. More often, history was seen as the causal result of social and economic forces alongside the religious instead of just explicit divine Providence. Though the treatment of God and Providence varied among Enlightenment thinkers, the secular was often promoted over the religious even by religious writers, who saw the hand of God as secondarily hidden behind the actions of men and institutions.

It is interesting that in a work of historiography that points out the primacy of white European thought and civilization in European Enlightenment historians’ writings, every historian mentioned is white and European. Was there no Enlightenment outside of Europe? And if the answer here is “no,” then what were comparable modes of historiographical thought outside of Europe, whether contemporary with the European Enlightenment or not? Surely, important developments in Enlightenment-era thought concerning historiography that shape the discipline today did not all come from Europe and European thinkers, alone. If they did, it reflects poorly on modern historiography to disregard or downplay such modes of doing history that do not come from a strictly European background. Of course, this is a brief survey of Enlightenment developments and only a few thinkers are introduced in the short chapter but I think it would benefit the overall picture immensely to hear something about Enlightenment historiography from a non-European perspective.